+++ zfec/fec.c
@@ -8,12 +8,8 @@
 #include <stdlib.h>
 #include <string.h>
 #include <assert.h>
+#include "tables.c"
 
-/*
- * Primitive polynomials - see Lin & Costello, Appendix A,
- * and  Lee & Messerschmitt, p. 453.
- */
-static const char*const Pp="101110001";
 
 
 /*
@@ -26,7 +22,6 @@ static const char*const Pp="101110001";
  */
 
 static gf gf_exp[510];  /* index->poly form conversion table    */
-static int gf_log[256]; /* Poly->index form conversion table    */
 static gf inverse[256]; /* inverse of field elem.               */
                                 /* inv[\alpha**i]=\alpha**(GF_SIZE-i-1) */
 
@@ -64,93 +59,9 @@ static gf gf_mul_table[256][256];
 #define GF_MULC0(c) __gf_mulc_ = gf_mul_table[c]
 #define GF_ADDMULC(dst, x) dst ^= __gf_mulc_[x]
 
-/*
- * Generate GF(2**m) from the irreducible polynomial p(X) in p[0]..p[m]
- * Lookup tables:
- *     index->polynomial form		gf_exp[] contains j= \alpha^i;
- *     polynomial form -> index form	gf_log[ j = \alpha^i ] = i
- * \alpha=x is the primitive element of GF(2^m)
- *
- * For efficiency, gf_exp[] has size 2*GF_SIZE, so that a simple
- * multiplication of two numbers can be resolved without calling modnn
- */
-static void
-_init_mul_table(void) {
-  int i, j;
-  for (i = 0; i < 256; i++)
-      for (j = 0; j < 256; j++)
-          gf_mul_table[i][j] = gf_exp[modnn (gf_log[i] + gf_log[j])];
-
-  for (j = 0; j < 256; j++)
-      gf_mul_table[0][j] = gf_mul_table[j][0] = 0;
-}
-
 #define NEW_GF_MATRIX(rows, cols) \
     (gf*)malloc(rows * cols)
 
-/*
- * initialize the data structures used for computations in GF.
- */
-static void
-generate_gf (void) {
-    int i;
-    gf mask;
-
-    mask = 1;                     /* x ** 0 = 1 */
-    gf_exp[8] = 0;          /* will be updated at the end of the 1st loop */
-    /*
-     * first, generate the (polynomial representation of) powers of \alpha,
-     * which are stored in gf_exp[i] = \alpha ** i .
-     * At the same time build gf_log[gf_exp[i]] = i .
-     * The first 8 powers are simply bits shifted to the left.
-     */
-    for (i = 0; i < 8; i++, mask <<= 1) {
-        gf_exp[i] = mask;
-        gf_log[gf_exp[i]] = i;
-        /*
-         * If Pp[i] == 1 then \alpha ** i occurs in poly-repr
-         * gf_exp[8] = \alpha ** 8
-         */
-        if (Pp[i] == '1')
-            gf_exp[8] ^= mask;
-    }
-    /*
-     * now gf_exp[8] = \alpha ** 8 is complete, so can also
-     * compute its inverse.
-     */
-    gf_log[gf_exp[8]] = 8;
-    /*
-     * Poly-repr of \alpha ** (i+1) is given by poly-repr of
-     * \alpha ** i shifted left one-bit and accounting for any
-     * \alpha ** 8 term that may occur when poly-repr of
-     * \alpha ** i is shifted.
-     */
-    mask = 1 << 7;
-    for (i = 9; i < 255; i++) {
-        if (gf_exp[i - 1] >= mask)
-            gf_exp[i] = gf_exp[8] ^ ((gf_exp[i - 1] ^ mask) << 1);
-        else
-            gf_exp[i] = gf_exp[i - 1] << 1;
-        gf_log[gf_exp[i]] = i;
-    }
-    /*
-     * log(0) is not defined, so use a special value
-     */
-    gf_log[0] = 255;
-    /* set the extended gf_exp values for fast multiply */
-    for (i = 0; i < 255; i++)
-        gf_exp[i + 255] = gf_exp[i];
-
-    /*
-     * again special cases. 0 has no inverse. This used to
-     * be initialized to 255, but it should make no difference
-     * since noone is supposed to read from here.
-     */
-    inverse[0] = 0;
-    inverse[1] = 1;
-    for (i = 2; i <= 255; i++)
-        inverse[i] = gf_exp[255 - gf_log[i]];
-}
 
 /*
  * Various linear algebra operations that i use often.
@@ -393,24 +304,6 @@ _invert_vdm (gf* src, unsigned k) {
     return;
 }
 
-/* There are few (if any) ordering guarantees that apply to reads and writes
- * of this static int across threads.  This is the reason for some of the
- * tight requirements for how `fec_init` is called.  If we could use a mutex
- * or a C11 atomic here we might be able to provide more flexibility to
- * callers.  It's tricky to do that while remaining compatible with all of
- * macOS/Linux/Windows and CPython's MSVC requirements and not switching to
- * C++ (or something even more different).
- */
-static int fec_initialized = 0;
-
-void
-fec_init (void) {
-    if (fec_initialized == 0) {
-        generate_gf();
-        _init_mul_table();
-        fec_initialized = 1;
-    }
-}
 
 /*
  * This section contains the proper FEC encoding/decoding routines.
@@ -438,10 +331,6 @@ fec_new(unsigned short k, unsigned short n) {
     assert(n >= 1);
     assert(n <= 256);
     assert(k <= n);
-
-    if (fec_initialized == 0) {
-        return NULL;
-    }
 
     retval = (fec_t *) malloc (sizeof (fec_t));
     retval->k = k;